feat: add yarn autoinstall script
