+++ .husky/post-checkout
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-node .husky/yarnLockCheck.js
+node .husky/yarnAutoInstall.js
